pull _.deepFilter generation into own module
